+++ infoschema/tables_test.go
@@ -1138,95 +1138,6 @@ select * from t3;
 	user2.MustQuery("select user,query from `CLUSTER_SLOW_QUERY` order by query").Check(testkit.Rows("user2 select * from t2;", "user2 select * from t3;"))
 }
 
-func (s *testClusterTableSuite) TestSlowQuery(c *C) {
-	writeFile := func(file string, data string) {
-		f, err := os.OpenFile(file, os.O_CREATE|os.O_WRONLY, 0644)
-		c.Assert(err, IsNil)
-		_, err = f.Write([]byte(data))
-		c.Assert(f.Close(), IsNil)
-		c.Assert(err, IsNil)
-	}
-
-	logData0 := ""
-	logData1 := `
-# Time: 2020-02-15T18:00:01.000000+08:00
-select 1;
-# Time: 2020-02-15T19:00:05.000000+08:00
-select 2;`
-	logData2 := `
-# Time: 2020-02-16T18:00:01.000000+08:00
-select 3;
-# Time: 2020-02-16T18:00:05.000000+08:00
-select 4;`
-	logData3 := `
-# Time: 2020-02-16T19:00:00.000000+08:00
-select 5;
-# Time: 2020-02-17T18:00:05.000000+08:00
-select 6;`
-	fileName0 := "tidb-slow-2020-02-14T19-04-05.01.log"
-	fileName1 := "tidb-slow-2020-02-15T19-04-05.01.log"
-	fileName2 := "tidb-slow-2020-02-16T19-04-05.01.log"
-	fileName3 := "tidb-slow.log"
-	writeFile(fileName0, logData0)
-	writeFile(fileName1, logData1)
-	writeFile(fileName2, logData2)
-	writeFile(fileName3, logData3)
-	defer func() {
-		os.Remove(fileName0)
-		os.Remove(fileName1)
-		os.Remove(fileName2)
-		os.Remove(fileName3)
-	}()
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	loc, err := time.LoadLocation("Asia/Shanghai")
-	c.Assert(err, IsNil)
-	tk.Se.GetSessionVars().TimeZone = loc
-	tk.MustExec("use information_schema")
-	cases := []struct {
-		sql    string
-		result []string
-	}{
-		{
-			sql:    "select count(*),min(time),max(time) from %s where time > '2019-01-26 21:51:00' and time < now()",
-			result: []string{"6|2020-02-15 18:00:01.000000|2020-02-17 18:00:05.000000"},
-		},
-		{
-			sql:    "select count(*),min(time),max(time) from %s where time > '2020-02-15 19:00:00' and time < '2020-02-16 18:00:02'",
-			result: []string{"2|2020-02-15 19:00:05.000000|2020-02-16 18:00:01.000000"},
-		},
-		{
-			sql:    "select count(*),min(time),max(time) from %s where time > '2020-02-16 18:00:02' and time < '2020-02-17 17:00:00'",
-			result: []string{"2|2020-02-16 18:00:05.000000|2020-02-16 19:00:00.000000"},
-		},
-		{
-			sql:    "select count(*),min(time),max(time) from %s where time > '2020-02-16 18:00:02' and time < '2020-02-17 20:00:00'",
-			result: []string{"3|2020-02-16 18:00:05.000000|2020-02-17 18:00:05.000000"},
-		},
-		{
-			sql:    "select count(*),min(time),max(time) from %s",
-			result: []string{"2|2020-02-16 19:00:00.000000|2020-02-17 18:00:05.000000"},
-		},
-		{
-			sql:    "select count(*),min(time) from %s where time > '2020-02-16 20:00:00'",
-			result: []string{"1|2020-02-17 18:00:05.000000"},
-		},
-		{
-			sql:    "select count(*) from %s where time > '2020-02-17 20:00:00'",
-			result: []string{"0"},
-		},
-		{
-			sql:    "select query from %s where time > '2019-01-26 21:51:00' and time < now()",
-			result: []string{"select 1;", "select 2;", "select 3;", "select 4;", "select 5;", "select 6;"},
-		},
-	}
-	for _, cas := range cases {
-		sql := fmt.Sprintf(cas.sql, "slow_query")
-		tk.MustQuery(sql).Check(testutil.RowsWithSep("|", cas.result...))
-		sql = fmt.Sprintf(cas.sql, "cluster_slow_query")
-		tk.MustQuery(sql).Check(testutil.RowsWithSep("|", cas.result...))
-	}
-}
-
 func (s *testTableSuite) TestSelectHiddenColumn(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("DROP DATABASE IF EXISTS `test_hidden`;")
